+++ config/database.js
@@ -1,7 +1,6 @@
 const mongoose = require('mongoose');
 
-// const URI = process.env.MONGO_DB_URI;
-const URI = "mongodb+srv://quick-call-team:fppxXhKPMcMQUKPu@quick-call.u7ccy.mongodb.net/quick-call?retryWrites=true&w=majority"
+const URI = process.env.MONGO_DB_URI;
 
 async function connectDB() {
     try {
